refactor(ai): consolidate treasury regiment-input test harness (#3941)
Extract shared game builders into treasury_planner_regiment_input_support
and merge six legacy shard files into bootstrap and supply/retention
contract suites with 1:1 row preservation (48 tests).
